Refactor: extract `Point3d`, `PointPad`, `Segment3d`, `SegmentPad` into `modmesh/universe/coord.hpp`
