+++ cpp/modmesh/universe/CMakeLists.txt
@@ -6,6 +6,7 @@ cmake_minimum_required(VERSION 3.16)
 set(MODMESH_UNIVERSE_HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/bernstein.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/bezier.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/coord.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/World.hpp
     CACHE FILEPATH "" FORCE)
