+++ cpp/modmesh/universe/universe.hpp
@@ -30,6 +30,7 @@
 
 #include <modmesh/universe/bernstein.hpp>
 #include <modmesh/universe/bezier.hpp>
+#include <modmesh/universe/coord.hpp>
 #include <modmesh/universe/World.hpp>
 
 namespace modmesh
